update readme and moved joshua_main
